+++ src/legacy/index.ts
@@ -1,2 +1,0 @@
-export { encryptCbc, decryptCbc } from "./cbc.js"
-export { openAuto } from "./auto.js"