+++ tsup.config.ts
@@ -3,7 +3,6 @@ import { defineConfig } from "tsup"
 export default defineConfig({
   entry: {
     index: "src/index.ts",
-    "legacy/index": "src/legacy/index.ts",
   },
   format: ["esm", "cjs"],
   dts: true,